fix(epics-base-rs): CALC engine C-parity (area 01)
Fixes parity-review area 01 (1 Critical, 7 High, 6 Medium, 5 Low):
- C-1/H-1: enforce end-of-expression stack-depth validation (final
  depth 0 store-terminated or 1; >1 rejected); synApps :=/UNTIL kept.
- H-2: max/min NaN propagation matches C accumulator semantics.
- H-3/H-4/H-5: nint/MODULO/bitwise use C d2i/d2ui 32-bit wrap.
- H-6/H-7: string evaluator uses exact IEEE compare and IEEE divide.
- M-3 hex literal 32-bit bound; M-5 CALC_ERR_* codes + calcErrorStr;
  M-6 cond_search count-based scan; L-2 calcArgUsage; L-3 disassemble;
  L-4 random range; L-5 FetchVal reads prev result.

+++ crates/epics-base-rs/src/calc/engine/token.rs
@@ -388,7 +388,9 @@ impl<'a> Tokenizer<'a> {
                 self.pos += 1;
             }
             let s = std::str::from_utf8(&self.input[start + 2..self.pos]).unwrap();
-            return u64::from_str_radix(s, 16)
+            // C postfix.c:283 parses hex literals via epicsParseUInt32 — a
+            // 32-bit unsigned value; anything wider is CALC_ERR_BAD_LITERAL.
+            return u32::from_str_radix(s, 16)
                 .map(|v| v as f64)
                 .map_err(|_| CalcError::BadLiteral);
         }
